add test free support barycenter algorithm + cleaning
